testbeds/gnmi/docker/network-controller-materialization/pyproject.toml
[tool.poetry]
name = "network-controller-materialization"
version = "1.0.0"
description = "gNMI Network Controller Materialization"
authors = ["GIROS-DIT UPM <gi.giros@upm.es>"]

[tool.poetry.dependencies]
python = "^3.9.5"
fastapi = {extras = ["all"], version = "^0.103.2"}
urllib3 = ">=1.25.3,<2.1.0"
markupsafe = ">=2.1.0,<2.1.5"
ngsi_ld_models = ">=1.0.0"
ngsi_ld_models_mdt_client_data_materialization = ">=1.0.0"
ngsi_ld_client = { git = "https://github.com/giros-dit/python-ngsi-ld-client.git", branch = "1.7.1" }
ncclient = "^0.6.13"
kafka-python = "^2.1.0"
python-dateutil = "^2.9.0"
jinja2 = "^3.1.4"
numpy = "^1.26.4"
ujson = ">=5.10.0"


[tool.poetry.group.dev.dependencies]
pytest = "^5.2"
flake8 = ">=4.0.1"
ngsi_ld_models = { path = "../../ngsi-ld-models", develop = true }
ngsi_ld_models_mdt_client_data_materialization = { path = "../../../ngsi-ld-models-mdt-client-data-materialization", develop = true }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
